% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riksdagen.R
\name{riksdagen}
\alias{riksdagen}
\title{THis is a description of riksdagen function}
\usage{
riksdagen(voteyear)
}
\arguments{
\item{voteyear}{a character with fixed format}
}
\value{
a data frame with percentage of voting data
}
\description{
riksdagen function creates an API to get data and makes some basic analysis.
}
\details{
The function creates an API connecting to website to get the data of voting in different parties and makes some basic analysis.
}
\examples{
 riksdagen("2020/21")

}
